pyproject.toml
[project]
name = "wh-task"
version = "0.1.0"
description = ""
authors = [
    {name = "jangil",email = "jangil.oh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "pytest (>=8.4.1,<9.0.0)",
    "allure-pytest (>=2.15.0,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "gradio (>=4.41.2,<5.0.0)",
    "langchain (>=0.2.11,<1.0.0)",
    "langgraph (>=0.1.1,<1.0.0)",
    "langchain-community (>=0.2.10,<1.0.0)",
    "langchain-core (>=0.2.23,<1.0.0)",
    "langchain-openai (>=0.1.18,<1.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "PyMuPDF (>=1.24.9,<2.0.0)",
    "langchain-google-genai (>=1.0.0,<2.0.0)",
    "pytest-mock (>=3.14.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "api_test" },
]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"

[tool.black]
line-length = 88
target-version = ['py312']

